boat_detector/requirements.txt
streamlit
torch
ultralytics
opencv-python
pillow
numpy
